reorganization for curve circuits
