+++ taiga_halo2/src/circuit/mod.rs
@@ -6,4 +6,5 @@ pub mod note_circuit;
 #[macro_use]
 pub mod vp_circuit;
 pub mod hash_to_curve;
+pub mod curve;
 pub mod vp_examples;
